requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0

# Image preprocessing
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
openai>=1.0.0
